# Cursor rules next phase 20250707 1031