--- a/rules/000-core/001-philosophers-stone.yaml
+++ b/rules/000-core/001-philosophers-stone.yaml
@@ -20,3 +20,5 @@ research: []
 examples: []
 notes: ''
 created: '2025-07-07'
+dependencies:
+- risk-checkpoint