--- a/rules/000-core/004-risk-checkpoint.mdc
+++ b/rules/000-core/004-risk-checkpoint.mdc
@@ -1,131 +1,137 @@
-# risk-checkpoint.mdc
+---
+description: "Final safety guardrail preventing high-risk operations"
+version: 2.0.0
+performance:
+  incident_reduction: 90-95%
+  assessment_latency: <100ms
+  token_budget: 650
+references:
+  - ISO31000:2018
+  - OWASP_Risk_Rating
+---
 
-<!-- Version: 1.0.0 — 2025-06-19 -->
-<!-- Based on ISO 31000:2018, OWASP Risk Rating, and AI Safety Best Practices -->
+# Risk Checkpoint
 
-* **Purpose**: To serve as the final, non-negotiable safety guardrail preventing execution of operations that carry unacceptable risk levels. This rule implements hierarchical risk assessment with context-aware thresholds, ensuring that agent capabilities never outpace safety mechanisms. **Empirical Evidence**: Systems with formal risk checkpoints show 90-95% critical incident reduction.
+**Purpose**: Block operations exceeding risk thresholds. 90-95% critical incident reduction.
 
-* **Requirements**:
-  * **Risk Classification Hierarchy**:
-    - **CRITICAL**: Data loss, security breach, system corruption
-    - **HIGH**: Service disruption, configuration damage, privacy violation  
-    - **MEDIUM**: Performance degradation, reversible changes, resource consumption
-    - **LOW**: Read operations, isolated changes, logging actions
-  * **Mandatory Checkpoints**:
-    - Pre-execution validation for ALL state-changing operations
-    - Pattern matching against risk signatures
-    - Sequence analysis for cumulative risk assessment
-    - Context evaluation (dev/staging/prod environment)
-  * **Risk Signatures** (Regex + Semantic): <!-- TODO: Expand pattern database -->
-    ```
-    CRITICAL:
-    - /rm\s+-rf\s+\// (recursive root deletion)
-    - /DROP\s+DATABASE|TRUNCATE/i (data destruction)
-    - /sudo\s+chmod\s+777/ (security compromise)
-    - API_KEY|SECRET|PASSWORD in outputs
-    - /.ssh/|/etc/passwd modifications
+## Risk Classification
+
+```python
+RISK_LEVELS = {
+    'CRITICAL': ['data_loss', 'security_breach', 'system_corruption'],
+    'HIGH': ['service_disruption', 'config_damage', 'privacy_violation'],
+    'MEDIUM': ['performance_impact', 'reversible_changes', 'resource_usage'],
+    'LOW': ['read_operations', 'isolated_changes', 'logging']
+}
+
+RISK_SIGNATURES = {
+    'CRITICAL': [
+        r'rm\s+-rf\s+/',                    # Root deletion
+        r'DROP\s+DATABASE|TRUNCATE',        # Data destruction
+        r'sudo\s+chmod\s+777',              # Security compromise
+        r'(API_KEY|SECRET|PASSWORD)',       # Credential exposure
+        r'(\.ssh/|/etc/passwd)',           # System file mods
+    ],
+    'HIGH': [
+        r'sudo(?!\s+apt-get)',             # Unwhitelisted sudo
+        r'DELETE.*WHERE.*[><=].*1000',     # Bulk operations
+        r'git\s+push.*main|master',        # Direct main push
+    ]
+}
+```
+
+## Risk Assessment Pipeline
+
+```python
+def assess_risk(operation):
+    # 1. Pattern matching
+    base_score = match_signatures(operation, RISK_SIGNATURES)
     
-    HIGH:
-    - /sudo/ without whitelist match
-    - Bulk operations (>100 items)
-    - External API calls with data payload
-    - Configuration file modifications
-    - Git push to main/master
-    ```
-  * **Dynamic Risk Scoring**:
-    - Base score from operation type
-    - Multipliers: frequency, data volume, user privileges
-    - Sequence detection: benign ops that combine dangerously
-  * **PFC-Inspired Conflict Detection**:
-    - Monitor for contradictory operations within same workflow
-    - Flag when current action conflicts with stated goals
-    - Detect decision-making loops indicating uncertainty
-    - **Empirical Benefit**: 91% precision in detecting decision conflicts
-  * **Executive Function Simulation**:
-    - Error monitoring: 31% false positive reduction
-    - Metacognitive assessment: "Does this action align with user intent?"
-    - Inhibitory control: Suppress impulsive high-risk operations
-    - Working memory: Track operation context across steps
-  * **Escalation Protocol**:
-    - LOW: Log and proceed
-    - MEDIUM: Warn and request confirmation
-    - HIGH: Require explicit approval with rationale
-    - CRITICAL: Block with mandatory human override
-
-* **Validation**:
-  * Check: Is risk level assessed BEFORE operation execution?
-  * Check: Are both individual and cumulative risks evaluated?
-  * Check: Do CRITICAL operations get blocked?
-  * Check: Is risk assessment logged for audit trail?
-  * Check: Are new operation types classified within 24 hours? <!-- TODO: Implement classification SLA -->
-
-* **Examples**:
-  * **Scenario**: Database migration script
-    ```
-    RISK ASSESSMENT:
-    Operation: "DROP TABLE users_backup; CREATE TABLE users_new..."
+    # 2. Context multipliers
+    multipliers = {
+        'environment': 2.0 if prod else 1.0,
+        'bulk_operation': 1.5 if count > 100 else 1.0,
+        'data_sensitivity': 2.0 if has_pii else 1.0,
+        'time_restriction': 1.5 if off_hours else 1.0
+    }
     
-    Individual Risks:
-    - DROP TABLE: HIGH (data loss potential)
-    - CREATE TABLE: LOW (additive operation)
+    # 3. Sequence analysis
+    if creates_dangerous_sequence(operation, history):
+        base_score = max(base_score, 'HIGH')
     
-    Sequence Risk: CRITICAL (dropping before verifying backup)
+    # 4. Final score
+    return apply_multipliers(base_score, multipliers)
+```
+
+## Decision Framework
+
+| Risk Level | Action | Requirements |
+|------------|--------|--------------|
+| **CRITICAL** | ❌ BLOCK | Human override + justification |
+| **HIGH** | ⚠️ WARN | Explicit approval + safeguards |
+| **MEDIUM** | 🔔 NOTIFY | Confirmation + logging |
+| **LOW** | ✅ PROCEED | Log for audit |
+
+## Implementation Patterns
+
+### Pre-execution Validation
+```python
+@risk_checkpoint
+def execute_operation(op):
+    risk = assess_risk(op)
     
-    CHECKPOINT RESULT: ❌ BLOCKED
-    Reason: Destructive operation without backup verification
-    Required: Confirm backup exists, add rollback plan
-    ```
-
-  * **Scenario**: Bulk file processing
-    ```
-    RISK ASSESSMENT:  
-    Operation: "Process 500 customer records, update status field"
+    if risk == 'CRITICAL':
+        raise BlockedOperation(f"Risk too high: {op.description}")
     
-    Context Evaluation:
-    - Environment: Production ⚠️
-    - Bulk threshold: >100 items ⚠️
-    - Operation type: Update (reversible) ✓
-    - Data sensitivity: PII present ⚠️
+    if risk == 'HIGH':
+        if not get_approval(op, required_safeguards):
+            raise ApprovalRequired(safeguards)
     
-    Cumulative Score: HIGH
-    
-    CHECKPOINT RESULT: ⚠️ APPROVAL REQUIRED
-    Safeguards: Require backup, batch in groups of 50, add rollback capability
-    ```
+    # Log and proceed
+    audit_log(op, risk)
+    return op.execute()
+```
 
-* **Implementation Architecture**:
-  ```
-  Request → Risk Parser → Pattern Match → Context Eval → Score → Decision
-                ↓              ↓              ↓           ↓         ↓
-            [Signatures]  [Risk DB]    [Environment] [Rules]  [Log/Block]
-  ```
+### Sequence Detection
+```python
+# Dangerous: DROP then CREATE without backup check
+sequence_patterns = [
+    ('DROP.*', 'CREATE.*', 'CRITICAL'),  # Missing backup
+    ('TRUNCATE', 'INSERT.*SELECT', 'HIGH'),  # Data replacement
+    ('UPDATE.*SET.*=', 'DELETE', 'HIGH'),  # Modify then delete
+]
+```
 
-* **Co-Evolution Mandate**:
-  * Every new agent capability MUST include:
-    - Risk pattern definitions
-    - Test cases for risk detection
-    - Mitigation strategies
-    - Update to risk signature database
-  * Pull requests adding features require risk assessment section
-  * Monthly review of "near miss" incidents to update patterns
+## Advanced Safeguards
 
-* **Advanced Safeguards**:
-  * **Honeypot Detection**: Fake resources that should never be accessed
-  * **Anomaly Detection**: Deviation from typical operation patterns
-  * **Time-based Restrictions**: Higher risk thresholds during off-hours
-  * **Rollback Preparation**: Automatic snapshot before HIGH risk ops
+- **Honeypots**: Fake resources triggering alerts
+- **Anomaly detection**: Deviation from baseline behavior
+- **Rollback prep**: Auto-snapshot before HIGH ops
+- **PFC simulation**: Conflict detection (91% precision)
 
-* **Performance Considerations**:
-  * Risk assessment must complete in <100ms for responsive UX
-  * Use bloom filters for initial pattern matching
-  * Cache risk scores for repeated operations
-  * Async logging to prevent blocking
+## Performance Optimization
 
-* **Changes**: Initial implementation v1.0.0. Establishes comprehensive risk management showing 90-95% critical incident reduction in systems with formal checkpoints.
+```python
+# Bloom filter for fast initial check
+bloom = BloomFilter(patterns=RISK_SIGNATURES)
+if not bloom.might_contain(operation):
+    return 'LOW'  # Fast path for safe ops
 
-* **Source References**: 
-  - ISO 31000:2018 Risk Management Guidelines
-  - OWASP Risk Rating Methodology
-  - AI Safety RICE Principles (Robustness, Interpretability, Controllability, Ethicality)
-  - Reuters AI Agent Risk Report (2024)
-  - https://pmc.ncbi.nlm.nih.gov/articles/PMC6850165/
+# Full assessment only for potential matches
+return detailed_assessment(operation)
+```
+
+## Co-evolution Mandate
+
+New capabilities MUST include:
+1. Risk pattern definitions
+2. Test cases for detection
+3. Mitigation strategies
+4. Update risk signature DB
+
+## Quick Validation
+- [ ] Risk assessed BEFORE execution
+- [ ] Individual + cumulative risks checked
+- [ ] CRITICAL ops blocked
+- [ ] Audit trail logged
+- [ ] <100ms assessment time