--- a/rules/100-cognitive/108-cognitive-load-balancing.mdc
+++ b/rules/100-cognitive/108-cognitive-load-balancing.mdc
@@ -98,62 +98,7 @@ Progress tracking: ▓░░░ 25%
 Cognitive load maintained in optimal zone: 0.45-0.62 ✓
 ```
 
-## Scientific Basis
 
-- Cognitive Load Theory (Sweller, 1988) demonstrates optimal learning at moderate load levels
-- Medical diagnostics study (2024): 42% cognitive load reduction with AI assistance, 29% accuracy improvement
-- Creative teams research: Peak performance at 0.45-0.62 CLT index
-- Working memory capacity limits (Miller's 7±2) require active load management
-- NASA Task Load Index (TLX) validated for human-AI interaction measurement
+## Extended Content
 
-## Trade-offs
-
-- **Pros**:
-  - Prevents cognitive overload and decision fatigue
-  - Maintains human engagement and skill development
-  - Optimizes overall system performance
-- **Cons**:
-  - Adds complexity to interaction design
-  - May slow down simple requests
-- **When to skip**:
-  - Emergency situations requiring all available processing
-  - Expert users who explicitly request full complexity
-  - Training scenarios designed to build capacity
-
-## Implementation Notes
-
-- Calculate CLT index using: (task_complexity × 0.4) + (presentation_complexity × 0.3) + (learning_required × 0.3)
-- Use progressive disclosure for complex information
-- Implement "cognitive load warnings" when approaching thresholds
-- Maintain load history to identify patterns and optimize over time
-- Consider individual differences - allow user-specific load preferences
-
-## Dependencies and Interactions
-
-- **Depends on**: Stepwise autonomy (for task breakdown)
-- **Enhances**: All cognitive rules through load awareness
-- **Conflicts with**: None, but modulates other rules' intensity
-
-## Metrics
-
-- **Measurement**: Average CLT index per session
-- **Target**: 80% of interactions in 0.45-0.62 range
-- **Method**: Calculate from response times, error rates, and task complexity
-
-## References
-
-- Sweller, J. (1988). Cognitive Load During Problem Solving
-- NASA TLX (Task Load Index) Assessment Tool
-- Chen et al. (2024). Optimal Cognitive Load Distribution in Human-AI Teams
-- Medical AI Study (2024): Cognitive Load Reduction in Diagnostic Tasks
-
-## Changelog
-
-- **1.0.0** - Initial version with CLT index monitoring
-
----
-*Rule metadata:*
-- *Author*: hamzaamjad
-- *Created*: 2025-01-02
-- *Last Updated*: 2025-01-02
-- *Stability*: experimental
+See notepad: `108-cognitive-load-balancing-extended.md`