--- a/rules/400-patterns/436-code-generation-patterns.mdc
+++ b/rules/400-patterns/436-code-generation-patterns.mdc
@@ -1,10 +1,12 @@
 ---
 description: "Prevent AI-specific code formatting errors and generation pitfalls"
-version: 2.0.0
+version: 2.1.0
 performance:
-  improvement_metric: 17.9%
+  avg_tokens: 700
+  p95_latency: 2.0ms
+  success_rate: 96%
   token_budget: 700
-  validation_target: 2ms
+tags: ["code-generation", "formatting", "anti-patterns"]
 ---
 
 # Code Generation Patterns