… .coverage
… .github/workflows/benchmark.yml
… .github/workflows/validate.yml
… CHANGELOG.md
… README.md
… baseline_metrics.json
… benchmark_report.json
… benchmark_results.json
… docs/generated/README.md
… docs/generated/categories/000-core.md
… docs/generated/categories/100-cognitive.md
… docs/generated/categories/200-domain.md
… docs/generated/categories/200-engineering.md
… docs/generated/categories/300-integration.md
… docs/generated/categories/300-techniques.md
… docs/generated/categories/400-patterns.md
… docs/generated/categories/500-safety.md
… docs/generated/categories/600-experimental.md
… docs/generated/categories/700-evolution.md
… docs/generated/categories/rules.md
… docs/generated/categories/symbiosis-engine.md
… docs/generated/images/dependency_graph.png
… docs/generated/rules/001-philosophers-stone.md
… docs/generated/rules/002-pareto-prioritization.md
… docs/generated/rules/003-stepwise-autonomy.md
… docs/generated/rules/004-risk-checkpoint.md
… docs/generated/rules/101-ultrathink-prompting.md
… docs/generated/rules/102-wildcard-brainstorm.md
… docs/generated/rules/103-divergence-convergence.md
… docs/generated/rules/104-analogy-transfer.md