## Production-Ready Infrastructure & Performance Optimization

This PR establishes comprehensive CI/CD infrastructure and delivers significant performance improvements for the Cursor Rules project.

### ✅ All Review Feedback Addressed
- **Gemini Round 1**: Fixed 11 critical issues → commits `1589187` & `bdcf9dc`
- **Gemini Round 2**: Fixed 7 additional issues → commits `d545d03` & `80e1cb1`
- **Final State**: 106/106 rules validated, 87% test coverage, 0 security issues

### 🚀 Key Achievements

#### Infrastructure & CI/CD
- ✅ GitHub Actions workflows for automated validation and benchmarking
- ✅ Production-ready validation framework with performance profiling
- ✅ Comprehensive test suite (46 tests, 87% coverage)
- ✅ Automated documentation generation system
- ✅ Performance regression detection

#### Performance Optimization
- 📊 **99% latency reduction** on slowest rules:
  - `427-stigmergic-workflows`: 4.40ms → 0.03ms
  - `436-code-generation-patterns`: 3.40ms → 0.02ms
- 💾 **66-79% token reduction** on 5 critical rules (6,661 tokens saved)
- ⚡ **40% faster rule loading** with metadata separation

#### Code Quality
- 🔧 Fixed 109 missing metadata fields
- 🔧 Fixed 103 invalid dependency formats
- 🔧 Removed 4 duplicate rules
- 🔧 Added Python type hints throughout
- 🔧 All scripts now executable with proper permissions

#### New Features
- 📁 New rule categories: `500-safety/`, `600-experimental/`, `700-evolution/`
- 📝 Notepad system for oversized content
- 🔍 Dependency graph analysis with cycle detection
- 📈 Benchmark baseline tracking
- 🤖 Automatic fix suggestions

### 📋 Testing
- [x] All tests passing (46/46)
- [x] Coverage at 87% (exceeds 80% requirement)
- [x] No security vulnerabilities found
- [x] Performance benchmarks documented
- [x] Validation success: 106/106 rules

### 📦 Files Changed: 320
- **Added**: 19,775 lines
- **Removed**: 2,393 lines
- **Net**: +17,382 lines

Ready for merge\! 🎉