@hamzaamjad 2025-07-07
## Review Feedback Addressed ✅

Thank you for the thorough review @gemini-code-assist! All issues have been resolved in commits 1589187 and bdcf9dc.

### Critical Issues Fixed:
1. **448-prompt-eval-extended.md** - Fixed corrupted syntax by removing invalid `},` prefix
2. **evolved-profiles.yaml** - Restored all 5 evolved profiles from git history (was critical data loss)
3. **fix_invalid_dependencies.py** - Completely rewrote to properly migrate old dependency format to new list format

### High Priority Issues Fixed:
4. **rule_validator.py:164** - Replaced bare except with `nx.NetworkXError`
5. **rule_validator.py:434** - Replaced bare except with `yaml.YAMLError`  
6. **rule_validator.py:291** - Fixed logic to read metadata from .yaml files instead of .mdc

### Medium Priority Issues Fixed:
7-11. Restored all comments and section headers to:
   - example_cursor_config.yaml
   - meta-rules-config.yaml
   - rule-config.yaml
   - move_oversized_to_notepad.py (now uses dynamic detection)

All validation tests still pass (106/106 rules). Ready for re-review.