@hamzaamjad 2025-07-07
## All Issues Resolved ✅

**Gemini Review (11/11):** Fixed
**Secondary Review (7/7):** Fixed
**Additional Fixes (4/4):** Complete

### Final State
```yaml
validation_success: 106/106
total_tokens: 99,305
dependencies_restored: 87
baseline_metrics: generated
security_issues: 0
```

### Commits
- `1589187` - Critical data loss fixes
- `bdcf9dc` - Comment restoration
- `d545d03` - Security hardening
- `80e1cb1` - Complete resolution

Ready for merge.